Moved some components around
